Moved the FirstRun infrastructure into a different folder so we can switch in any other databases easily.
