+++ MCS-Extractor/ImportedData/Postgres/FirstRunWindow.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="MCS_Extractor.FirstRunWindow"
+﻿<Window x:Class="MCS_Extractor.FirstRun.Postgres.PostgresFirstRunWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:MCS_Extractor"
+        xmlns:local="clr-namespace:MCS_Extractor.FirstRun.Postgres"
         mc:Ignorable="d"
         Title="First Run" Height="300" Width="500">
     <Grid>
